requirements.txt
bs4>=0.0.2
google-api-python-client>=2.170.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.2
ipykernel>=6.29.5
langchain>=0.3.25
langchain-groq>=0.3.2
langgraph>=0.4.5
pydantic>=2.11.4
pymongo>=4.13.0
python-dotenv>=1.1.0
rebuff>=0.0.5